MAINT: merge torch implementations of _flips.py to _flips
