+++ torch_np/_detail/__init__.py
@@ -3,4 +3,3 @@ from ._reductions import *
 
 # leading underscore (ndarray.flatten yes, np.flatten no)
 from .implementations import *
-from .implementations import _flatten
